Loan post impl done

+++ src/main/java/com/abdullah/pms/repository/GradeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.abdullah.pms.domain.Grade;
 
-public interface GradeRepository extends JpaRepository<Grade, Long>{
+public interface GradeRepository extends JpaRepository<Grade, Integer>{
 
 }
