+++ src/main/java/com/abdullah/pms/repository/ShiftRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.abdullah.pms.domain.Shift;
 
-public interface ShiftRepository extends JpaRepository<Shift, Long>{
+public interface ShiftRepository extends JpaRepository<Shift, Integer>{
 
 }
